+++ vsc-base/out/vsc-base-vscode-webview.js
@@ -169,15 +169,15 @@ exports.initWebview = ({ viewType = 'vscScript', title = 'Script', html, body, s
  * This method returns two async methods:\
  * 'postMessage' which post a message from the extension to the webview \
  * and 'createdOnMessage' which creates a awaited receiver for messages send from the webview. \
- * 'createdOnMessage' take a onMessage call back with two arguments: (message: any) and (dispose: ()=>void) \
- * The 'createdOnMessage' will await until the dispose method is called, which will stop the webview, \
+ * 'createdOnMessage' take a onMessage call back with two arguments: (message: any) and (resolve: ()=>void) \
+ * The 'createdOnMessage' will await until the resolve method is called, \
  * and continue code after. (Normally it will end the execution of the extension there) \
  * See [startWebview](http://vsc-base.org/#startWebview) for full detail and example.
  * @see [setupWebviewConnection](http://vsc-base.org/#setupWebviewConnection)
  * @internal
  * @vscType webview
  * @returns [postMessage, createdOnMessage]
- * @returns [(message: any) => Promise<boolean>, (callback: (message: any, dispose: () => void) => void) => Promise<void>]
+ * @returns [(message: any) => Promise<boolean>, (callback: (message: any, resolve: () => void) => void) => Promise<void>]
  * @example
  * const [postMessage, createdOnMessage] = vsc.setupWebviewConnection(context, webviewPanel)
  */
@@ -192,23 +192,25 @@ exports.setupWebviewConnection = (context, webviewPanel) => {
         return false;
     });
     const proxy = {
-        onMessage: (message, dispose) => { },
-        dispose: (value) => { }
+        onMessage: (message, resolve) => { },
+        resolve: (value) => { }
     };
     webviewPanel.webview.onDidReceiveMessage(message => proxy.onMessage &&
-        proxy.onMessage(message, proxy.dispose), undefined, context.subscriptions);
+        proxy.onMessage(message, proxy.resolve), undefined, context.subscriptions);
+    const dispose = () => {
+        if (webviewPanel) {
+            webviewPanel.dispose();
+        }
+    };
     const createdOnMessage = (onMessage) => __awaiter(this, void 0, void 0, function* () {
         proxy.onMessage = onMessage;
         return new Promise((resolve) => {
-            proxy.dispose = () => {
-                if (webviewPanel) {
-                    webviewPanel.dispose();
-                }
+            proxy.resolve = () => {
                 resolve();
             };
         });
     });
-    return [postMessage, createdOnMessage, webviewPanel.dispose, webviewPanel];
+    return [postMessage, createdOnMessage, dispose, webviewPanel];
 };
 /** vsc-base method
  * @description
@@ -246,13 +248,13 @@ exports.setupWebviewConnection = (context, webviewPanel) => {
  *      }
  *    }
  *  });
- *  await onMessage(async (message, dispose) => {
+ *  await onMessage(async (message, resolve) => {
  *    switch (message.command) {
  *      case "show":
  *        vsc.showMessage(message.value);
  *        break;
  *      case "end":
- *        dispose();
+ *        resolve();
  *        break;
  *      case "search":
  *        const files = await vsc.findFilePaths(message.value);
@@ -260,6 +262,7 @@ exports.setupWebviewConnection = (context, webviewPanel) => {
  *        break;
  *    }
  *  });
+ *  dispose();
  *  vsc.showMessage('Done!')
  */
 exports.startWebview = (context, startOptions) => {