+++ vsc-script/src/vsc-base-development/vsc-base-vscode-webview.ts
@@ -178,15 +178,15 @@ export const initWebview = ({
  * This method returns two async methods:\
  * 'postMessage' which post a message from the extension to the webview \
  * and 'createdOnMessage' which creates a awaited receiver for messages send from the webview. \
- * 'createdOnMessage' take a onMessage call back with two arguments: (message: any) and (dispose: ()=>void) \
- * The 'createdOnMessage' will await until the dispose method is called, which will stop the webview, \
+ * 'createdOnMessage' take a onMessage call back with two arguments: (message: any) and (resolve: ()=>void) \
+ * The 'createdOnMessage' will await until the resolve method is called, \
  * and continue code after. (Normally it will end the execution of the extension there) \
  * See [startWebview](http://vsc-base.org/#startWebview) for full detail and example.
  * @see [setupWebviewConnection](http://vsc-base.org/#setupWebviewConnection)
  * @internal
  * @vscType webview
  * @returns [postMessage, createdOnMessage]
- * @returns [(message: any) => Promise<boolean>, (callback: (message: any, dispose: () => void) => void) => Promise<void>]
+ * @returns [(message: any) => Promise<boolean>, (callback: (message: any, resolve: () => void) => void) => Promise<void>]
  * @example
  * const [postMessage, createdOnMessage] = vsc.setupWebviewConnection(context, webviewPanel)
  */
@@ -204,32 +204,34 @@ export const setupWebviewConnection = (
       return false
    }
    const proxy: {
-      onMessage: (message: any, dispose: () => void) => void
-      dispose: (value?: unknown) => void
+      onMessage: (message: any, resolve: () => void) => void
+      resolve: (value?: unknown) => void
    } = {
-      onMessage: (message, dispose) => { },
-      dispose: (value) => { }
+      onMessage: (message, resolve) => { },
+      resolve: (value) => { }
    }
    webviewPanel.webview.onDidReceiveMessage(
       message =>
          proxy.onMessage &&
-         proxy.onMessage(message, proxy.dispose)
+         proxy.onMessage(message, proxy.resolve)
       ,
       undefined,
       context.subscriptions
    );
-   const createdOnMessage = async (onMessage: (message: any, dispose: () => void) => void): Promise<void> => {
+   const dispose = () => {
+      if (webviewPanel) {
+         webviewPanel.dispose();
+      }
+   }
+   const createdOnMessage = async (onMessage: (message: any, resolve: () => void) => void): Promise<void> => {
       proxy.onMessage = onMessage;
       return new Promise((resolve) => {
-         proxy.dispose = () => {
-            if (webviewPanel) {
-               webviewPanel.dispose();
-            }
+         proxy.resolve = () => {
             resolve()
          }
       })
    }
-   return [postMessage, createdOnMessage, webviewPanel.dispose, webviewPanel]
+   return [postMessage, createdOnMessage, dispose, webviewPanel]
 }
 
 /** vsc-base method
@@ -268,13 +270,13 @@ export const setupWebviewConnection = (
  *      }
  *    }
  *  });
- *  await onMessage(async (message, dispose) => {
+ *  await onMessage(async (message, resolve) => {
  *    switch (message.command) {
  *      case "show":
  *        vsc.showMessage(message.value);
  *        break;
  *      case "end":
- *        dispose();
+ *        resolve();
  *        break;
  *      case "search":
  *        const files = await vsc.findFilePaths(message.value);
@@ -282,6 +284,7 @@ export const setupWebviewConnection = (
  *        break;
  *    }
  *  });
+ *  dispose();
  *  vsc.showMessage('Done!')
  */
 export const startWebview = (context: vscode.ExtensionContext, startOptions: vsc.IStartWebviewOptions): vsc.WebviewConnectionReturnType => {