+++ vsc-script/src/webview-test1.vsc-script.ts
@@ -3,7 +3,7 @@ import * as vsc from "vsc-base";
 import * as vscode from "vscode";
 
 export async function run(path: string, context: vscode.ExtensionContext) {
-   const [postMessage, onMessage] = vsc.startWebview(context, {
+   const [postMessage, onMessage, dispose] = vsc.startWebview(context, {
       title: "Rename",
       body: `
       <button onClick="postMessage({command:'show',value:'1'})">Show '1'</button>
@@ -21,15 +21,20 @@ export async function run(path: string, context: vscode.ExtensionContext) {
          }
       }
    });
-   await onMessage(async (message, dispose) => {
+   await onMessage(async (message, resolve) => {
       switch (message.command) {
-         case "show": vsc.showMessage(message.value); break;
-         case "end": dispose(); break;
+         case "show":
+            vsc.showMessage(message.value);
+            break;
+         case "end":
+            resolve();
+            break;
          case "search":
             const files = await vsc.findFilePaths(message.value);
             postMessage({ command: "info", content: files.length });
             break;
       }
    });
+   dispose()
    vsc.showMessage("Script Done!");
 }